Split out styles into their own file
